wxWidgets v2.6.2 incremental update
git-svn-id: http://svn.osafoundation.org/chandler/trunk@6303 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe

+++ internal/wx/src/mac/carbon/dccg.cpp
@@ -1882,7 +1882,7 @@ bool wxDC::DoGetPartialTextExtents(const wxString& text, wxArrayInt& widths) con
     status = ::ATSUCreateTextLayoutWithTextPtr( (UniCharArrayPtr) ubuf , 0 , chars , chars , 1 ,
         &chars , (ATSUStyle*) &m_macATSUIStyle , &atsuLayout ) ;
         
-    	for ( int pos = 0; pos < (int)chars; pos ++ ) {
+    	for ( int pos = 0; pos < chars; pos ++ ) {
 			unsigned long actualNumberOfBounds = 0;
 			ATSTrapezoid glyphBounds;
